Swithc to 32 bit

+++ src/client/mod.rs
@@ -32,15 +32,15 @@ pub struct StravaClient {
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Activity {
-    pub id: u64,
+    pub id: u32,
     pub name: String,
-    pub distance: f64,
-    pub moving_time: u64,
-    pub elapsed_time: u64,
-    pub total_elevation_gain: f64,
+    pub distance: f32,
+    pub moving_time: u32,
+    pub elapsed_time: u32,
+    pub total_elevation_gain: f32,
     pub sport_type: String,
-    pub average_heartrate: Option<f64>,
-    pub max_heartrate: Option<f64>,
+    pub average_heartrate: Option<f32>,
+    pub max_heartrate: Option<f32>,
     pub start_date: DateTime<Utc>,
 }
